// frontend/src/types/auth.types.ts
export interface Sede {
  id: number;
  name: string;
  address: string;
  phone: string;
}

export interface User {
  id: number;
  email: string;
  first_name: string;
  last_name: string;
  full_name: string;
  phone: string;
  role: 'ADMINISTRATOR' | 'ENCARGADO' | 'WORKER' | 'CASHIER' | 'CUSTOMER';
  sede: Sede | null;
  is_active: boolean;
  created_at: string;
  // Security fields
  login_attempts:   number;
  locked_until:     string | null;
  unlock_requested: boolean;
}